Rebuit the UI using tripleplay 1.4
Added badges and logging of game state.

+++ core/src/main/java/au/edu/unimelb/csse/trollsvsgoats/core/model/Animation.java
@@ -9,7 +9,7 @@ public class Animation {
     // A parent image with subimages.
     private Image parentImage;
     private Image[] images;
-    private int frameWidth;
+    private float frameWidth;
     private int frameCount;
     private float frameTime;
     private int frameIndex;
